修改user表和video表结构体的id为int64

+++ model/user.go
@@ -2,9 +2,9 @@ package model
 
 //User_info:用户信息
 type User_info struct {
-	ID             int `gorm:"column:id;primary_key;AUTO_INCREMENT;not null"`
+	ID             int64 `gorm:"column:id;primary_key;AUTO_INCREMENT;not null"`
 	Name           string
-	Follow_count   int
-	Follower_count int
+	Follow_count   int64
+	Follower_count int64
 	Password_Hash  string // 加密后的密码
 }
